requirements.txt
openai-agents>=0.1.0
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
cognitora>=1.6.1

